Move App files into frontend folder, remove unused old files
