+++ src/frontend/App.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Route, Switch, withRouter } from 'react-router-dom';
-import logo from './logo.svg';
 import './App.css';
 
 function App() {
