+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css';
-import App from './App';
+import App from './frontend/App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(<Router><App /></Router>, document.getElementById('root'));
